+++ frontend/src/app/features/fermentation/fermentation-dashboard.component.ts
@@ -7,7 +7,7 @@ import { debounceTime } from 'rxjs';
 import { ApiService, DataFilters, FermentationRequest, FermentationResponse } from '../../core/services/api.service';
 import { AuthService } from '../../core/services/auth.service';
 import { ToastService } from '../../core/services/toast.service';
-import { FermentationGraphPanelComponent } from '../../shared/components/fermentation-graphs/fermentation-yield-graph/fermentation-yield-graph.component';
+import { FermentationGraphPanelComponent } from './fermentation-graph-panel/fermentation-graph-panel.component';
 
 
 type ModalFieldKey =
