+++ frontend/src/app/features/fermentation/fermentation-graph-panel/fermentation-graph-panel.component.css
